kore-python/requirements.txt
kaggle-environments
IPython
python-logstash
python-logstash-async